refactor: split the observer's IO adapter and the tab-snapshot read tests
`activity-observer.ts` held both the pure loop (cadence, tracks, edges, claims,
every fact injected) and the production adapter that opens sockets, reads
`pty-exits.json` and writes the registry. Those are two jobs; the adapter moves
to `activity-observer-io.ts` and imports the loop's contract type, so the
direction is one-way and no value-import cycle can form.

`api-tab-snapshot.test.ts` was already at 483 lines on main and mixed writers
that need a temp HOME with joins that touch no filesystem. The pure half moves
to `api-tab-join.test.ts`.
